Fix issue with uint32_t index

+++ lib/ethash/ethash-internal.hpp
@@ -21,7 +21,7 @@ hash256 calculate_seed(uint32_t epoch_number) noexcept;
 
 light_cache make_light_cache(size_t size, const hash256& seed);
 
-hash512 calculate_full_dataset_item(const light_cache& cache, uint32_t index);
+hash512 calculate_full_dataset_item(const light_cache& cache, size_t index);
 
 hash256 hash_light(const epoch_context& context, const hash256& header_hash, uint64_t nonce);
